Added tag_key support :
A tag key is a file following this pattern :

#delete#.reso1.reso2.reso3.except.quality1.quality2

You can use as much resolution and quality that you want, the .except.qual1.qual2.qualn... is optional.
in this case file being in reso1, reso2 and reso3 will be move to the unwanted directorie, except if
their quality is equal to 1 or 2

Not that is is impossible actualy to "except" different quality for different resolution.

available resolution :
1080p, 720p, 560p, 480p, unknown
available quality :
web, bdrip, bluray, dvdrip, hdrip, hdtv, unknown

This is not case sensitive